Refactor colliders (#17)
* added in object axes and updated colliders accordingly

* fixed some of the colliders and created basic demo of collision detection

* made the input system a singleton

* refactored code

* some bug fixes

* optimization for input manager

* bug fixes

* optimization for normalization

* added a GetSize() function for box colliders

* some cleanup on gjk algorithm

* fixed bug with rigidbodies and fixing up convex shape colliders

* convex shape colliders should be working now

* added some functionality to the collision detection demo

* bug fixes to rigidbodies and demo is done

* fixed a mini issue with the demo

* some modification

* some mini changes

+++ MTRXEngine/MTRXEngine/source/code/GameTime.cpp
@@ -47,7 +47,7 @@ namespace mtrx
 	// Gets what time it is right now
 	float GameTime::GetTime()
 	{
-		return glfwGetTime();
+		return (float)glfwGetTime();
 	}
 
 	// Calculate delta time
